+++ RELEASE_NOTES.md
@@ -3,3 +3,5 @@ v0.71.0
 - Java, Node, Swift: changed IdentityKeyStore.saveIdentity to return an enum.
 
 - Node: RegistrationService.registerAccount takes account password as a string.
+
+- keytrans: Bridge to Node